requirements.txt
flask
flask-cors
bcrypt
python-dotenv
pandas
numpy
scikit-learn
flask-jwt-extended
requests
gunicorn
yfinance
